All examples compile again but untested...

+++ examples/spectrum/spectrum-eval.c
@@ -44,7 +44,7 @@ int main(int argc, char **argv) {
     fprintf(stderr, "Can't open spectrum.dat for writing.\n");
     exit(1);
   }
-  for (iter = 0, en = -0.5 * spectrum->step * spectrum->nx; iter < spectrum->nx; iter++, en += spectrum->step)
+  for (iter = 0, en = -0.5 * spectrum->step * (REAL) spectrum->nx; iter < spectrum->nx; iter++, en += spectrum->step)
     fprintf(fp, FMT_R " " FMT_R " " FMT_R "\n", en, CREAL(cgrid_value_at_index(spectrum, 1, 1, iter)), CIMAG(cgrid_value_at_index(spectrum, 1, 1, iter)));
   fclose(fp);
   printf("Spectrum written to spectrum.dat\n");
